fixed tabs on bottom, added code back from xmoon
I did a diff of my fork and the original, and I found out that some of the stuff I removed didn't actually need to be removed, so I readded some of that stuff. The navbar stuff is staying gone, as that's what makes the tabs unified, but I'll re-add those lines if there's a way to do that in the darwin-specific browser.css, but I don't know if that's possible.