make unified titlebar mac-only, readd navbar stuff
Updated /os/darwin/browser.css to include the unified titlebar stuff from the default theme, works perfectly. Should be cross-platform but I'm not sure, so I'll try to replicate the changes I've made using the original source, and hopefully it should all work without issue, but the only way to truly test is to get this to work cross-platform is to test it on all platforms.